AG-10755 - factor our bar group scale calculations
