+++ packages/ag-charts-enterprise/src/series/candlestick/candlestickThemes.ts
@@ -4,7 +4,6 @@ export const CANDLESTICK_SERIES_THEME = {
     series: {
         __extends__: _Theme.EXTENDS_SERIES_DEFAULTS,
         direction: 'vertical' as const,
-        strokeWidth: 2,
     },
     axes: {
         [_Theme.CARTESIAN_AXIS_TYPE.NUMBER]: {
